authentication/requirements.txt
flask
flask_mysqldb
flask-bcrypt
python-dotenv
flask-session
flask-cors
requests